weeds

+++ common/src/main/java/com/ordana/immersive_weathering/data/block_growths/growths/ConfigurableBlockGrowth.java
@@ -229,10 +229,8 @@ public class ConfigurableBlockGrowth implements IBlockGrowth {
 
                         if (areaCondition.test(pos, level, this)) {
 
-                            if (destroyTarget) {
-                                level.destroyBlock(targetPos, true);
-                            }
-                            level.setBlockAndUpdate(targetPos, getStateToPlace(toPlace.getFirst(), target, level, pos));
+                            if (destroyTarget) level.destroyBlock(targetPos, true);
+                            level.setBlock(targetPos, getStateToPlace(toPlace.getFirst(), target, level, pos), 2);
                             if (db) {
                                 if (destroyTarget) level.destroyBlock(targetPos2, true);
                                 level.setBlockAndUpdate(targetPos2, getStateToPlace(toPlace.getSecond(), target2, level, pos));
